requirements.txt

pandas
scikit-learn
joblib
kafka-python
pyspark
